hickory_proto

Enum ProtoErrorKind

#[non_exhaustive]
pub enum ProtoErrorKind {

    BadTransactionId,
    Busy,

    #[non_exhaustive]
    CharacterDataTooLong {
        max: usize,
        len: usize,
    },
    Decode(DecodeError),
    Dns(DnsError),
    FormError {
        header: Header,
        error: Box<ProtoError>,
    },
    MaxBufferSizeExceeded(usize),
    Message(&'static str),
    Msg(String),
    NoConnections,

    #[non_exhaustive]
    NotAllRecordsWritten {
        count: usize,
    },
    Io(Arc<Error>),
    Timeout,
    UrlParsing(ParseError),
    Utf8(Utf8Error),
    FromUtf8(FromUtf8Error),
    ParseInt(ParseIntError),
    QueryCaseMismatch,

}

The error kind for errors that get returned in the crate

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

BadTransactionId

A UDP response was received with an incorrect transaction id, likely indicating a cache-poisoning attempt.

Busy

The underlying resource is too busy

This is a signal that an internal resource is too busy. The intended action should be tried again, ideally after waiting for a little while for the situation to improve. Alternatively, the action could be tried on another resource (for example, in a name server pool).

#[non_exhaustive]
CharacterDataTooLong

Character data length exceeded the limit

Fields

Non-exhaustive enum variants could have additional fields added in future. Therefore, non-exhaustive enum variants cannot be constructed in external crates and cannot be matched against.

max: usize

Specified maximum

len: usize

Actual length

Decode(DecodeError)

Message decoding error

Dns(DnsError)

Semantic DNS errors

FormError

Format error in Message Parsing

Fields

header: Header

Header of the bad Message

error: Box<ProtoError>

Error that occurred while parsing the Message

MaxBufferSizeExceeded(usize)

The maximum buffer size was exceeded

Message(&'static str)

An error with an arbitrary message, referenced as &’static str

Msg(String)

An error with an arbitrary message, stored as String

NoConnections

No resolvers available

#[non_exhaustive]
NotAllRecordsWritten

Not all records were able to be written

Fields

Non-exhaustive enum variants could have additional fields added in future. Therefore, non-exhaustive enum variants cannot be constructed in external crates and cannot be matched against.

count: usize

Number of records that were written before the error

Io(Arc<Error>)

An error got returned from IO

Timeout

A request timed out

UrlParsing(ParseError)

An url parsing error

Utf8(Utf8Error)

A utf8 parsing error

FromUtf8(FromUtf8Error)

A utf8 parsing error

ParseInt(ParseIntError)

An int parsing error

QueryCaseMismatch

Case randomization is enabled, and a server did not echo a query name back with the same case.

Implementations

impl ProtoErrorKind

pub fn is_bad_transaction_id(&self) -> bool

Returns true if this is a ProtoErrorKind::BadTransactionId, otherwise false

pub fn is_busy(&self) -> bool

Returns true if this is a ProtoErrorKind::Busy, otherwise false

pub fn is_character_data_too_long(&self) -> bool

Returns true if this is a ProtoErrorKind::CharacterDataTooLong, otherwise false

pub fn as_character_data_too_long_mut( &mut self, ) -> Option<(&mut usize, &mut usize)>

Optionally returns mutable references to the inner fields if this is a ProtoErrorKind::CharacterDataTooLong, otherwise None

pub fn as_character_data_too_long(&self) -> Option<(&usize, &usize)>

Optionally returns references to the inner fields if this is a ProtoErrorKind::CharacterDataTooLong, otherwise None

pub fn into_character_data_too_long(self) -> Result<(usize, usize), Self>

Returns the inner fields if this is a ProtoErrorKind::CharacterDataTooLong, otherwise returns back the enum in the Err case of the result

pub fn is_decode(&self) -> bool

Returns true if this is a ProtoErrorKind::Decode, otherwise false

pub fn as_decode_mut(&mut self) -> Option<&mut DecodeError>

Optionally returns mutable references to the inner fields if this is a ProtoErrorKind::Decode, otherwise None

pub fn as_decode(&self) -> Option<&DecodeError>

Optionally returns references to the inner fields if this is a ProtoErrorKind::Decode, otherwise None

pub fn into_decode(self) -> Result<DecodeError, Self>

Returns the inner fields if this is a ProtoErrorKind::Decode, otherwise returns back the enum in the Err case of the result

pub fn is_dns(&self) -> bool

Returns true if this is a ProtoErrorKind::Dns, otherwise false

pub fn as_dns_mut(&mut self) -> Option<&mut DnsError>

Optionally returns mutable references to the inner fields if this is a ProtoErrorKind::Dns, otherwise None

pub fn as_dns(&self) -> Option<&DnsError>

Optionally returns references to the inner fields if this is a ProtoErrorKind::Dns, otherwise None

pub fn into_dns(self) -> Result<DnsError, Self>

Returns the inner fields if this is a ProtoErrorKind::Dns, otherwise returns back the enum in the Err case of the result

pub fn is_form_error(&self) -> bool

Returns true if this is a ProtoErrorKind::FormError, otherwise false

pub fn as_form_error_mut( &mut self, ) -> Option<(&mut Header, &mut Box<ProtoError>)>

Optionally returns mutable references to the inner fields if this is a ProtoErrorKind::FormError, otherwise None

pub fn as_form_error(&self) -> Option<(&Header, &Box<ProtoError>)>

Optionally returns references to the inner fields if this is a ProtoErrorKind::FormError, otherwise None

pub fn into_form_error(self) -> Result<(Header, Box<ProtoError>), Self>

Returns the inner fields if this is a ProtoErrorKind::FormError, otherwise returns back the enum in the Err case of the result

pub fn is_max_buffer_size_exceeded(&self) -> bool

Returns true if this is a ProtoErrorKind::MaxBufferSizeExceeded, otherwise false

pub fn as_max_buffer_size_exceeded_mut(&mut self) -> Option<&mut usize>

Optionally returns mutable references to the inner fields if this is a ProtoErrorKind::MaxBufferSizeExceeded, otherwise None

pub fn as_max_buffer_size_exceeded(&self) -> Option<&usize>

Optionally returns references to the inner fields if this is a ProtoErrorKind::MaxBufferSizeExceeded, otherwise None

pub fn into_max_buffer_size_exceeded(self) -> Result<usize, Self>

Returns the inner fields if this is a ProtoErrorKind::MaxBufferSizeExceeded, otherwise returns back the enum in the Err case of the result

pub fn is_message(&self) -> bool

Returns true if this is a ProtoErrorKind::Message, otherwise false

pub fn as_message_mut(&mut self) -> Option<&mut &'static str>

Optionally returns mutable references to the inner fields if this is a ProtoErrorKind::Message, otherwise None

pub fn as_message(&self) -> Option<&&'static str>

Optionally returns references to the inner fields if this is a ProtoErrorKind::Message, otherwise None

pub fn into_message(self) -> Result<&'static str, Self>

Returns the inner fields if this is a ProtoErrorKind::Message, otherwise returns back the enum in the Err case of the result

pub fn is_msg(&self) -> bool

Returns true if this is a ProtoErrorKind::Msg, otherwise false

pub fn as_msg_mut(&mut self) -> Option<&mut String>

Optionally returns mutable references to the inner fields if this is a ProtoErrorKind::Msg, otherwise None

pub fn as_msg(&self) -> Option<&String>

Optionally returns references to the inner fields if this is a ProtoErrorKind::Msg, otherwise None

pub fn into_msg(self) -> Result<String, Self>

Returns the inner fields if this is a ProtoErrorKind::Msg, otherwise returns back the enum in the Err case of the result

pub fn is_no_connections(&self) -> bool

Returns true if this is a ProtoErrorKind::NoConnections, otherwise false

pub fn is_not_all_records_written(&self) -> bool

Returns true if this is a ProtoErrorKind::NotAllRecordsWritten, otherwise false

pub fn as_not_all_records_written_mut(&mut self) -> Option<&mut usize>

Optionally returns mutable references to the inner fields if this is a ProtoErrorKind::NotAllRecordsWritten, otherwise None

pub fn as_not_all_records_written(&self) -> Option<&usize>

Optionally returns references to the inner fields if this is a ProtoErrorKind::NotAllRecordsWritten, otherwise None

pub fn into_not_all_records_written(self) -> Result<usize, Self>

Returns the inner fields if this is a ProtoErrorKind::NotAllRecordsWritten, otherwise returns back the enum in the Err case of the result

pub fn is_io(&self) -> bool

Returns true if this is a ProtoErrorKind::Io, otherwise false

pub fn as_io_mut(&mut self) -> Option<&mut Arc<Error>>

Optionally returns mutable references to the inner fields if this is a ProtoErrorKind::Io, otherwise None

pub fn as_io(&self) -> Option<&Arc<Error>>

Optionally returns references to the inner fields if this is a ProtoErrorKind::Io, otherwise None

pub fn into_io(self) -> Result<Arc<Error>, Self>

Returns the inner fields if this is a ProtoErrorKind::Io, otherwise returns back the enum in the Err case of the result

pub fn is_timeout(&self) -> bool

Returns true if this is a ProtoErrorKind::Timeout, otherwise false

pub fn is_url_parsing(&self) -> bool

Returns true if this is a ProtoErrorKind::UrlParsing, otherwise false

pub fn as_url_parsing_mut(&mut self) -> Option<&mut ParseError>

Optionally returns mutable references to the inner fields if this is a ProtoErrorKind::UrlParsing, otherwise None

pub fn as_url_parsing(&self) -> Option<&ParseError>

Optionally returns references to the inner fields if this is a ProtoErrorKind::UrlParsing, otherwise None

pub fn into_url_parsing(self) -> Result<ParseError, Self>

Returns the inner fields if this is a ProtoErrorKind::UrlParsing, otherwise returns back the enum in the Err case of the result

pub fn is_utf8(&self) -> bool

Returns true if this is a ProtoErrorKind::Utf8, otherwise false

pub fn as_utf8_mut(&mut self) -> Option<&mut Utf8Error>

Optionally returns mutable references to the inner fields if this is a ProtoErrorKind::Utf8, otherwise None

pub fn as_utf8(&self) -> Option<&Utf8Error>

Optionally returns references to the inner fields if this is a ProtoErrorKind::Utf8, otherwise None

pub fn into_utf8(self) -> Result<Utf8Error, Self>

Returns the inner fields if this is a ProtoErrorKind::Utf8, otherwise returns back the enum in the Err case of the result

pub fn is_from_utf8(&self) -> bool

Returns true if this is a ProtoErrorKind::FromUtf8, otherwise false

pub fn as_from_utf8_mut(&mut self) -> Option<&mut FromUtf8Error>

Optionally returns mutable references to the inner fields if this is a ProtoErrorKind::FromUtf8, otherwise None

pub fn as_from_utf8(&self) -> Option<&FromUtf8Error>

Optionally returns references to the inner fields if this is a ProtoErrorKind::FromUtf8, otherwise None

pub fn into_from_utf8(self) -> Result<FromUtf8Error, Self>

Returns the inner fields if this is a ProtoErrorKind::FromUtf8, otherwise returns back the enum in the Err case of the result

pub fn is_parse_int(&self) -> bool

Returns true if this is a ProtoErrorKind::ParseInt, otherwise false

pub fn as_parse_int_mut(&mut self) -> Option<&mut ParseIntError>

Optionally returns mutable references to the inner fields if this is a ProtoErrorKind::ParseInt, otherwise None

pub fn as_parse_int(&self) -> Option<&ParseIntError>

Optionally returns references to the inner fields if this is a ProtoErrorKind::ParseInt, otherwise None

pub fn into_parse_int(self) -> Result<ParseIntError, Self>

Returns the inner fields if this is a ProtoErrorKind::ParseInt, otherwise returns back the enum in the Err case of the result

pub fn is_query_case_mismatch(&self) -> bool

Returns true if this is a ProtoErrorKind::QueryCaseMismatch, otherwise false