Module hickory_proto::error

Error types for the crate

Structs

DnsSecError

The error type for dnssec errors that get returned in the crate

ProtoError

The error type for errors that get returned in the crate

Enums

DnsSecErrorKind

The error kind for dnssec errors that get returned in the crate

ProtoErrorKind

The error kind for errors that get returned in the crate

Traits

FromProtoError

A trait marking a type which implements From and std::error::Error types as well as Clone + Send

Type Aliases

DnsSecResult

An alias for dnssec results returned by functions of this crate

ProtoResult

An alias for results returned by functions of this crate