hickory_proto

Module op

Operations to send with a Client or server, e.g. Query, Message, or UpdateMessage can be used together to either query or update resource records sets.

Re-exports

pub use update_message::UpdateMessage;

Modules

update_message

Update related operations for Messages

Structs

DnsRequest

A DNS request object

DnsRequestOptions

A set of options for expressing options to how requests should be treated

DnsResponse

A DNS response object

Edns

Edns implements the higher level concepts for working with extended dns as it is used to create or be created from OPT record data.

EdnsFlags

EDNS flags

Header

Metadata for the Message struct.

LowerQuery

Identical to Query, except that the Name is guaranteed to be in lower case form.

Message

The basic request and response data structure, used for all DNS protocols.

MessageParts

Consumes Message giving public access to fields in Message so they can be destructured and taken by value

Query

Query struct for looking up resource records, basically a resource record without RDATA.

SerialMessage

A DNS message in serialized form, with either the target address or source address

Enums

MessageSignature

Indicates how a Message is signed.

MessageType

Message types are either Query (also Update) or Response

OpCode

Operation code for queries, updates, and responses

ResponseCode

The status code of the response to a query.

Traits

EmitAndCount

A trait that defines types which can be emitted as a set, with the associated count returned.

MessageSigner

A trait for adding a final MessageSignature to a Message before it is sent.

ResponseSigner

A trait for producing a MessageSignature for responses

Functions

emit_message_parts

Emits the different sections of a message properly

Type Aliases

MessageVerifier

Alias for a function verifying if a message is properly signed