Struct hickory_proto::op::Edns

pub struct Edns { /* private fields */ }

Edns implements the higher level concepts for working with extended dns as it is used to create or be created from OPT record data.

Implementations

impl Edns

pub fn new() -> Self

Creates a new extended DNS object.

pub fn rcode_high(&self) -> u8

The high order bytes for the response code in the DNS Message

pub fn version(&self) -> u8

Returns the EDNS version

pub fn dnssec_ok(&self) -> bool

Specifies that DNSSEC is supported for this Client or Server

pub fn max_payload(&self) -> u16

Maximum supported size of the DNS payload

pub fn option(&self, code: EdnsCode) -> Option<&EdnsOption>

Returns the Option associated with the code

pub fn options(&self) -> &OPT

Returns the options portion of EDNS

pub fn options_mut(&mut self) -> &mut OPT

Returns a mutable options portion of EDNS

pub fn set_rcode_high(&mut self, rcode_high: u8) -> &mut Self

Set the high order bits for the result code.

pub fn set_version(&mut self, version: u8) -> &mut Self

Set the EDNS version

pub fn set_dnssec_ok(&mut self, dnssec_ok: bool) -> &mut Self

Set to true if DNSSEC is supported

pub fn set_max_payload(&mut self, max_payload: u16) -> &mut Self

Set the maximum payload which can be supported From RFC 6891: Values lower than 512 MUST be treated as equal to 512

pub fn set_option(&mut self, option: EdnsOption)

๐Deprecated: Please use options_mut().insert() to modify

Set the specified EDNS option