hickory_proto::op

Trait MessageSigner

pub trait MessageSigner:
    Send
    + Sync
    + 'static {
    // Required method
    fn sign_message(
        &self,
        message: &Message,
        current_time: u64,
    ) -> Result<(MessageSignature, Option<MessageVerifier>), ProtoError>;

    // Provided method
    fn should_sign_message(&self, message: &Message) -> bool { ... }
}

A trait for adding a final MessageSignature to a Message before it is sent.

Required Methods

fn sign_message( &self, message: &Message, current_time: u64, ) -> Result<(MessageSignature, Option<MessageVerifier>), ProtoError>

Finalize the provided Message, computing a MessageSignature, and optionally providing a MessageVerifier for response messages.

Arguments

message - the message to finalize

current_time - the current system time.

Return

A MessageSignature to append to the end of the additional data, and optionally a MessageVerifier to use to verify responses provoked by the message.

Provided Methods

fn should_sign_message(&self, message: &Message) -> bool

Return whether the message requires a signature before being sent. By default, returns true for AXFR and IXFR queries, and Update and Notify messages