hickory_proto::op

Module update_message

Update related operations for Messages

Constants

MAX_PAYLOAD_LEN

Maximum payload length for EDNS update messages

Traits

UpdateMessage

To reduce errors in using the Message struct as an Update, this will do the call throughs to properly do that.

Functions

append

Appends a record to an existing rrset, optionally require the rrset to exist (atomicity depends on the server)

compare_and_swap

Compares and if it matches, swaps it for the new value (atomicity depends on the server)

create

Sends a record to create on the server, this will fail if the record exists (atomicity depends on the server)

delete_all

Deletes all records at the specified name

delete_by_rdata

Deletes a record (by rdata) from an rrset, optionally require the rrset to exist.

delete_rrset

Deletes an entire rrset, optionally require the rrset to exist.

zone_transfer

Download all records from a zone, or all records modified since given SOA was observed. The request will either be a AXFR Query (ask for full zone transfer) if a SOA was not provided, or a IXFR Query (incremental zone transfer) if a SOA was provided.