hickory_proto::op::update_message

Function delete_all

pub fn delete_all(
    name_of_records: Name,
    zone_origin: Name,
    dns_class: DNSClass,
    use_edns: bool,
) -> Message

Deletes all rrsets at the specified name

RFC 2136, DNS Update, April 1997

2.5.3 - Delete All RRsets From A Name

  One RR is added to the Update Section whose NAME is that of the name
  to be cleansed of RRsets.  TYPE must be specified as ANY.  TTL must
  be specified as zero (0) and is otherwise not used by the Primary
  Zone Server.  CLASS must be specified as ANY.  RDLENGTH must be zero (0)
  and RDATA must therefore be empty.  If no such RRsets exist, then
  this Update RR will be silently ignored by the Primary Zone Server.

Arguments

name_of_records - the name of all the record sets to delete

zone_origin - the zone name to update, i.e. SOA name

dns_class - the class of the SOA

use_edns - if true, EDNS options will be added to the request

The update must go to a zone authority (i.e. the server used in the ClientConnection). This operation attempts to delete all resource record sets at the specified name, regardless of the record type.