hickory_proto::op::update_message

Function delete_by_rdata

pub fn delete_by_rdata(
    rrset: RecordSet,
    zone_origin: Name,
    use_edns: bool,
) -> Message

Deletes a record (by rdata) from an rrset

RFC 2136, DNS Update, April 1997

2.5.4 - Delete An RR From An RRset

  RRs to be deleted are added to the Update Section.  The NAME, TYPE,
  RDLENGTH and RDATA must match the RR being deleted.  TTL must be
  specified as zero (0) and will otherwise be ignored by the Primary
  Zone Server.  CLASS must be specified as NONE to distinguish this from an
  RR addition.  If no such RRs exist, then this Update RR will be
  silently ignored by the Primary Zone Server.

Arguments

rrset - the record(s) to delete from a RRSet, the name, type and rdata must match the record to delete

zone_origin - the zone name to update, i.e. SOA name

use_edns - if true, EDNS options will be added to the request

The update must go to a zone authority (i.e. the server used in the ClientConnection). If no such RRs exist, then this Update RR will be silently ignored by the Primary Zone Server.