hickory_proto::rr

Module rdata

All record data structures and related serialization methods

Re-exports

pub use self::a::A;

pub use self::aaaa::AAAA;

pub use self::caa::CAA;

pub use self::cert::CERT;

pub use self::csync::CSYNC;

pub use self::hinfo::HINFO;

pub use self::https::HTTPS;

pub use self::mx::MX;

pub use self::name::ANAME;

pub use self::name::CNAME;

pub use self::name::NS;

pub use self::name::PTR;

pub use self::naptr::NAPTR;

pub use self::null::NULL;

pub use self::openpgpkey::OPENPGPKEY;

pub use self::opt::OPT;

pub use self::soa::SOA;

pub use self::srv::SRV;

pub use self::sshfp::SSHFP;

pub use self::svcb::SVCB;

pub use self::tlsa::TLSA;

pub use self::txt::TXT;

Modules

a

IPv4 address record data

aaaa

IPv6 address record data

caa

allows a DNS domain name holder to specify one or more Certification Authorities (CAs) authorized to issue certificates for that domain.

cert

CERT record type for storing certificates in DNS

csync

CSYNC record for synchronizing data from a child zone to the parent

hinfo

HINFO record for storing host information

https

HTTPS type and related implementations

mx

mail exchange, email, record

name

Record type for all cname like records.

naptr

Dynamic Delegation Discovery System

null

null record type, generally not used except as an internal tool for representing null data

openpgpkey

OPENPGPKEY records for OpenPGP public keys

opt

option record for passing protocol options between the client and server

soa

start of authority record defining ownership and defaults for the zone

srv

service records for identify port mapping for specific services on a host

sshfp

SSHFP records for SSH public key fingerprints

svcb

SVCB records, see RFC 9460 SVCB and HTTPS Resource Records, Nov 2023

tlsa

TLSA records for storing TLS certificate validation information

txt

text records for storing arbitrary data