Function hickory_proto::rr::rdata::caa::read_iodef

pub fn read_iodef(url: &[u8]) -> ProtoResult<Url>

Incident Object Description Exchange Format

RFC 8659, DNS Certification Authority Authorization, November 2019

5.4.  CAA iodef Property

   The iodef property specifies a means of reporting certificate issue
   requests or cases of certificate issue for the corresponding domain
   that violate the security policy of the issuer or the domain name
   holder.

   The Incident Object Description Exchange Format (IODEF) [RFC7970](https://www.rfc-editor.org/info/rfc7970) is
   used to present the incident report in machine-readable form.

   The iodef property takes a URL as its parameter.  The URL scheme type
   determines the method used for reporting:

   mailto:  The IODEF incident report is reported as a MIME email
      attachment to an SMTP email that is submitted to the mail address
      specified.  The mail message sent SHOULD contain a brief text
      message to alert the recipient to the nature of the attachment.

   http or https:  The IODEF report is submitted as a Web service
      request to the HTTP address specified using the protocol specified
      in [RFC6546].