hickory_proto

Module runtime

Abstractions to deal with different async runtimes.

Structs

TokioHandle

A handle to the Tokio runtime

TokioRuntimeProvider

The Tokio Runtime for async execution

TokioTime

New type which is implemented using tokio::time::{Delay, Timeout}

Traits

Executor

Generic executor.

QuicSocketBinder

Noop trait for when the quinn dependency is not available.

RuntimeProvider

RuntimeProvider defines which async runtime that handles IO and timers.

Spawn

A type defines the Handle which can spawn future.

Time

Generic Time for Delay and Timeout.

Functions

spawn_bg

Spawn a background task, if it was present