hickory_proto

Trait Time

pub trait Time {
    // Required methods
    fn delay_for<'async_trait>(
        duration: Duration,
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>;
    fn timeout<'async_trait, F>(
        duration: Duration,
        future: F,
    ) -> Pin<Box<dyn Future<Output = Result<F::Output, Error>> + Send + 'async_trait>>
       where F: 'async_trait + 'static + Future + Send;
}

Generic Time for Delay and Timeout.

Required Methods

fn delay_for<'async_trait>( duration: Duration, ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>

Return a type that implements Future that will wait until the specified duration has elapsed.

fn timeout<'async_trait, F>( duration: Duration, future: F, ) -> Pin<Box<dyn Future<Output = Result<F::Output, Error>> + Send + 'async_trait>>
where F: 'async_trait + 'static + Future + Send,

Return a type that implement Future to complete before the specified duration has elapsed.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.