Crate hickory_proto

Hickory DNS Protocol library

Modules

op

Operations to send with a Client or server, e.g. Query, Message, or UpdateMessage can be used together to either query or update resource records sets.

rr

Resource record related components, e.g. Name aka label, Record, RData, …

runtime

Abstractions to deal with different async runtimes.

serialize

Contains serialization libraries for binary and text, txt.

tcp

TCP protocol related components for DNS

udp

UDP protocol related components for DNS

xfer

DNS high level transit implementations.

Structs

ForwardData

Data needed to process a SOA-record-based referral.

ForwardNSData

Data needed to process a NS-record-based referral.

ProtoError

The error type for errors that get returned in the crate

Enums

ProtoErrorKind

The error kind for errors that get returned in the crate