hickory_proto::op

Module message

Basic protocol message for DNS

Structs

HeaderCounts

Tracks the counts of the records in the Message.

Message

The basic request and response data structure, used for all DNS protocols.

MessageParts

Consumes Message giving public access to fields in Message so they can be destructured and taken by value

Traits

EmitAndCount

A trait that defines types which can be emitted as a set, with the associated count returned.

MessageFinalizer

A trait for performing final amendments to a Message before it is sent.

Functions

count_was_truncated

Returns the count written and a boolean if it was truncated

emit_message_parts

Emits the different sections of a message properly

update_header_counts

Returns a new Header with accurate counts for each Message section

Type Aliases

MessageVerifier

Alias for a function verifying if a message is properly signed