hickory_proto::op::message

Trait EmitAndCount

pub trait EmitAndCount {
    // Required method
    fn emit(
        &mut self,
        encoder: &mut BinEncoder<'_>,
    ) -> Result<usize, ProtoError>;
}

A trait that defines types which can be emitted as a set, with the associated count returned.

Required Methods

fn emit(&mut self, encoder: &mut BinEncoder<'_>) -> Result<usize, ProtoError>

Emit self to the encoder and return the count of items