Module hickory_proto::rr::rdata::caa

allows a DNS domain name holder to specify one or more Certification Authorities (CAs) authorized to issue certificates for that domain.

RFC 8659, DNS Certification Authority Authorization, November 2019

The Certification Authority Authorization (CAA) DNS Resource Record
allows a DNS domain name holder to specify one or more Certification
Authorities (CAs) authorized to issue certificates for that domain.
CAA Resource Records allow a public Certification Authority to
implement additional controls to reduce the risk of unintended
certificate mis-issue.  This document defines the syntax of the CAA
record and rules for processing CAA records by certificate issuers.

Structs

CAA

The CAA RR Type

KeyValue

Issuer key and value pairs.

Enums

Property

Specifies in what contexts this key may be trusted for use

Value

Potential values.

Functions

read_iodef

Incident Object Description Exchange Format

read_issuer

Reads the issuer field according to the spec