hickory_proto::rr::rdata

Module name

Record type for all cname like records.

A generic struct for all {*}NAME pointer RData records, CNAME, NS, and PTR. Here is the text for CNAME from RFC 1035, Domain Implementation and Specification, November 1987:

RFC 1035, DOMAIN NAMES - IMPLEMENTATION AND SPECIFICATION, November 1987

3.3.1. CNAME RDATA format

    +--+--+--+--+--+--+--+--+--+--+--+--+--+--+--+--+
    /                     CNAME                     /
    /                                               /
    +--+--+--+--+--+--+--+--+--+--+--+--+--+--+--+--+

where:

CNAME           A <domain-name> which specifies the canonical or primary
                name for the owner.  The owner name is an alias.

CNAME RRs cause no additional section processing, but name servers may
choose to restart the query at the canonical name in certain cases.  See
the description of name server logic in [RFC-1034] for details.

Structs

ANAME

new type for the RecordData of ANAME

CNAME

new type for the RecordData of CNAME

NS

new type for the RecordData of NS

PTR

new type for the RecordData of PTR

Functions

emit

RFC 4034, DNSSEC Resource Records, March 2005

read

Read the RData from the given Decoder